Refactored FrameBuffer and RenderPass into a single thing (called RenderPass).

+++ Humble2/src/Humble2/Resources/ResourceManager.cpp
@@ -30,7 +30,7 @@ namespace HBL2
 		return m_DeletionQueue;
 	}
 
-	void ResourceManager::Flush(uint32_t currentFrame)
+	void ResourceManager::Flush(uint64_t currentFrame)
 	{
 		m_DeletionQueue.Flush(currentFrame);
 	}
